kubernetes/wp-k8s-service/requirements.txt
fastapi==0.109.0
pydantic==2.5.3
uvicorn[standard]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
python-multipart==0.0.6
lxml==5.1.0
boto3==1.34.34
playwright==1.40.0
playwright-stealth==1.0.6
loguru==0.7.2
camoufox[geoip,zip]==0.3.5
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-exporter-otlp-proto-http==1.22.0
kubernetes==29.0.0
pymysql==1.1.0
dramatiq[redis,watch]==1.15.0
redis==5.0.1
aioredis==2.0.1
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
